moved TerraWDIOSpecReporter from src and cleaned the code
