+++ docker-compose.yml
@@ -12,7 +12,7 @@ services:
       TZ: 'America/Chicago'
 
   dev:
-    image: 'cerner/terra-node-dev:1'
+    image: cerner/terra-node-dev:1
     ports:
       - 8080:8080
     depends_on:
