+++ wdio.conf.js
@@ -3,7 +3,7 @@ const fs = require('fs');
 const wdioConf = require('./config/wdio/wdio.conf');
 const webpackConfigObject = require('./tests/test.config.js');
 const webpackConfigFunction = require('./tests/test.config.func.js');
-const WdioFileReporter = require('./src/wdio/WdioFileReporter');
+const WdioFileReporter = require('./reporters/wdio/TerraWDIOSpecReporter');
 
 WdioFileReporter.reporterName = 'TT_WDIO_CR';
